Extract `StringSerachAlgorithm`
